Add a feature-flagged `--ignore-local` flag to `hab pkg install`
Following the base plans refresh of June 2018, we discovered a small
bug in how our package installation logic could resolve dependencies
that are part of the Studio. In particular, this can occur when using
the `HAB_BLDR_CHANNEL` environment variable to specify an
other-than-stable channel to pull packages from.

There is logic in the Rust code that will end up using a
locally-installed package if something cannot be found in the
specified channel in Builder. If this is the case, the plan build
software will not fall back to look in the `stable` channel. This
would be an issue for us when building the next Habitat release,
because the first parts of our release are built with the
currently-stable Habitat, which was built before the base plans
refresh. Because of how we build using a release channel, we would end
up pulling in some dependencies based on what is installed in the
Studio, rather than allowing the build logic to fall back to pull from
stable. We can then run into issues later in the release process
building later packages that do end up pulling in newer dependencies;
we end up with multiple versions and thus cannot build.

For the next release, we will have these dependencies pinned
explicitly to ensure we get pre-refresh packages.

That brings us to this commit.

This introduces an `--ignore-local` flag for `hab pkg install` that
disables this "fallback to a local package" logic. Once this is
present in a stable `hab` release, we'll be able to build using
packages exclusively from Builder, as `hab-plan-build` has been
modified to use it.

This issue probably only shows up when you "straddle" a base plans
refresh like this.

Additionally, there may be better ways to ultimately handle this,
which is why this capability is currently behind a feature-flag. To
use it, you must set the environment variable `HAB_FEAT_IGNORE_LOCAL`
to `true`, which makes the `--ignore-local` option available on the
`hab pkg install` command. As such, this behavior *IS NOT* guaranteed
to remain in the `hab` CLI, pending a solidification of the desired
behavior.

Signed-off-by: Christopher Maier <cmaier@chef.io>